requirements.txt
# === Core ===
python-dotenv>=1.0.1
pydantic>=2.9.2

# === LLM / Orchestration ===
groq>=0.11.0
google-generativeai>=0.3.2
langchain>=0.3.0
langgraph>=0.1.15

# === API / Server ===
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-multipart>=0.0.9
slowapi>=0.1.9

# === Database (Supabase / PostgreSQL) ===
supabase>=2.10.0

# === Utils / HTTP ===
requests>=2.32.3

# === OSINT / Scraping ===
beautifulsoup4>=4.12.3
